Added py files to separate the different methods by their category.
